requirements.txt
Django==2.2.17
djangorestframework==3.9.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.6.0
djangorestframework-jwt==1.11.0
jsonpath==0.82